// tangent-cc-lib/src/lib/type/font-logo.type.ts
export enum FontLogo {
  Ubuntu = '\uf31b',
  ArchLinux = '\uf303',
  CentOS = '\uf304',
  Debian = '\uf306',
  Elementary = '\uf309',
  Fedora = '\uf30a',
  FreeBSD = '\uf30c',
  Gentoo = '\uf30d',
  Mageia = '\uf310',
  Mandriva = '\uf311',
  Manjaro = '\uf312',
  LinuxMint = '\uf30e',
  OpenBSD = '\uf328',
  RaspberryPi = '\uf315',
  RedHat = '\uf316',
  Sabayon = '\uf317',
  Tux = '\uf31a',
}